Tab data loading on nav clicks
Set up a simple switch statement in both cases (clicking on the images
in the welcome tab, and clicking on the actual nav tabs) that can be
added to with relative ease as more tabs are developed.  These
statements will be responsible for setting the pulled data from the
database into the forms that are loaded when navigating between tabs.

Removed an extra </div>

Fixed indentation.

Added ecoAjax API to the head